Complete test for 'right.applier' function factory
